requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
numpy>=1.24.0

